Refactoring (2)
